requirements.txt
Flask==2.3.3
mysql-connector-python==8.1.0
Werkzeug==2.3.7
requests==2.31.0
python-dotenv==1.0.0
Flask-WTF==1.1.1
WTForms==3.0.1
email-validator==2.0.0